% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baseline_funds}
\alias{baseline_funds}
\title{Budget balance on a know start date}
\format{
A data frame with 17 rows and 3 variables:
\describe{
\item{date}{date of budget balance, character expressed in YYYY-MM-DD}
\item{amount}{balance of budget balance, in US dollars}
\item{category}{budget category, character}
...
}
}
\usage{
baseline_funds
}
\description{
Balances for each budget category on a defined date. These balances are
calculated in the notebook \strong{internal_data.Rmd} in the \strong{data-raw}
folder.
}
\keyword{datasets}
