% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{budget_balance}
\alias{budget_balance}
\title{Budget balances on May 15, 2021}
\format{
A data frame with 17 rows and 3 variables:
\describe{
\item{category}{budget category, character}
\item{balance}{balance of budget balance, in US dollars}
...
}
}
\usage{
budget_balance
}
\description{
Budget balances on May 15, 2021.
}
\keyword{datasets}
