% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{budget_defaults}
\alias{budget_defaults}
\title{Default budget monthly levels}
\format{
A data frame with 17 rows and 2 variables:
\describe{
\item{category}{budget category, character}
\item{amount}{balance of budget balance, in US dollars}
...
}
}
\usage{
budget_defaults
}
\description{
Default monthly values.
}
\keyword{datasets}
