% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{budget_monthly_wide}
\alias{budget_monthly_wide}
\title{Monthly budget levels}
\format{
A data frame with 18 variables:
\describe{
\item{date}{date of budget allotment, with the full date representing
the month that date is in, character expressed in YYYY-MM-DD}
\item{routine_maintenance}{Monthly budget allotment for the Routine
Maintenance category, in US dollars}
\item{service_parts}{Monthly budget allotment for the Service Parts
category, in US dollars}
\item{entertainment}{Monthly budget allotment for the Entertainment
category, in US dollars}
\item{donation}{Monthly budget allotment for the Donation
category, in US dollars}
\item{gift}{Monthly budget allotment for the Gift
category, in US dollars}
\item{running_events}{Monthly budget allotment for the Running Events
category, in US dollars}
\item{home_goods_large}{Monthly budget allotment for the Home Goods -
Large category, in US dollars}
\item{home_goods_small}{Monthly budget allotment for the Home Goods -
Small category, in US dollars}
\item{home_supplies}{Monthly budget allotment for the Home Supplies
category, in US dollars}
\item{flex_spending}{Monthly budget allotment for the Flex Spending
category, in US dollars}
\item{personal_care}{Monthly budget allotment for the Personal Care
category, in US dollars}
\item{athletic_apparel}{Monthly budget allotment for the Athletic
Apparel category, in US dollars}
\item{bike_maintenance}{Monthly budget allotment for the Bike Maintenance
category, in US dollars}
\item{electronics_software}{Monthly budget allotment for the Electronics
& Software category, in US dollars}
\item{sporting_goods}{Monthly budget allotment for the Sporting Goods
category, in US dollars}
\item{fashion_apparel}{Monthly budget allotment for the Fashion
Apparel category, in US dollars}
\item{travel}{Monthly budget allotment for the Travel category, in
US dollars}
...
}
}
\usage{
budget_monthly_wide
}
\description{
The amount allocated to every budget category, recorded on a monthly basis.
}
\keyword{datasets}
