% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{category_key}
\alias{category_key}
\title{Category label key}
\format{
A data frame with 17 rows and 2 variables:
\describe{
\item{category}{budget category label, character}
\item{category_clean}{budget category label processed with
\strong{janitor::make_clean_names()}, character}
...
}
}
\usage{
category_key
}
\description{
Look-up table for connecting the budget category labels with same labels
processed by the \strong{janitor::make_clean_names()} function.
}
\keyword{datasets}
