% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{future_transactions}
\alias{future_transactions}
\title{Future transactions}
\format{
A data frame with 3 variables:
\describe{
\item{date}{date of budget balance, character expressed in YYYY-MM-DD}
\item{amount}{balance of budget balance, in US dollars}
\item{category}{budget category, character}
...
}
}
\usage{
future_transactions
}
\description{
Anticipated future transactions.
}
\keyword{datasets}
